more warnings cleaned up
git-svn-id: svn+ssh://svn.pointclouds.org/pcl/trunk@4990 a9d63959-f2ad-4865-b262-bf0e56cfafb6

+++ registration/include/pcl/registration/impl/ia_ransac.hpp
@@ -191,7 +191,7 @@ pcl::SampleConsensusInitialAlignment<PointSource, PointTarget, FeatureT>::comput
 
   if (!error_functor_)
   {
-    error_functor_.reset (new TruncatedError (corr_dist_threshold_));
+    error_functor_.reset (new TruncatedError (static_cast<float> (corr_dist_threshold_)));
   }
 
   std::vector<int> sample_indices (nr_samples_);
@@ -201,9 +201,10 @@ pcl::SampleConsensusInitialAlignment<PointSource, PointTarget, FeatureT>::comput
 
   final_transformation_ = guess;
   int i_iter = 0;
-  if (!guess.isApprox(Eigen::Matrix4f::Identity(), 0.01f)) { //If guess is not the Identity matrix we check it.
+  if (!guess.isApprox(Eigen::Matrix4f::Identity (), 0.01f)) 
+  { //If guess is not the Identity matrix we check it.
 	  transformPointCloud (*input_, input_transformed, final_transformation_);
-	  lowest_error = computeErrorMetric (input_transformed, (float) corr_dist_threshold_);
+	  lowest_error = computeErrorMetric (input_transformed, static_cast<float> (corr_dist_threshold_));
 	  i_iter = 1;
   }
